refactor: Generate data type union instead of hand writing
